[yykamei] ステップ16: 複数人で利用できるようにしよう（ユーザの導入） (#959)
* Add User

`User` is responsible for storing **User**'s information: `name` and
`password`.

So far, authentication has not implemented yet. This patch just adds a
new model and applied some validations to it.

* Create association between Task and User

This associates `Task` with `User`.
For the next migration step, this also adds
the task named `data_migration:associate_tasks_to_tester`, which
updates all old `Task`s to belong to `tester`. This is rquired to make
`user_id` NOT NULL.

* Change user_id NOT NULL

* Add bullet