Add Label
`Label` is responsible fore storing **Label**s to be attached to
**Task**s. `Label` and `Task` will have the relationship as Many to
Many, but this patch does not include the implementation for now.

Note `Label` is owned by a user, and the user can't own a label named
the same as the one's other labels. However, another user can name a
label as the same as the other's label.