Add :dependent to has_many on User, Task, and Label
If an instance is deleted, the associated objects should be deleted,
too. Note `dependent: :destory` can cause significant performance
problem because each related record is deleted one by one. This training
program does not care about such a problem. In the first place, we can
skip the callback with `#delete_all`, so it's OK to use it if the
problem appears.